Remove the views when they are not needed
Fixes #3

Android only blocks sensitive actions, when an app is currently drawing
over the screen. By removing the views when they are not needed, users
will rarely be inconvenienced by this, since the icon is ussualy not in
view. If it is in view and they try to install something, simply waiting
till the icon dissapears will fix their issue.